+++ src/ska_sdp_dataproduct_api/components/store/persistent/postgresql.py
@@ -28,22 +28,46 @@ class PostgresConnector(MetadataStore):
     A class to connect to a PostgreSQL instance and test its availability.
     """
 
-    def __init__(self, host: str, port: int, user: str, password: str, table_name: str):
+    def __init__(
+        self, host: str, port: int, user: str, password: str, schema: str, table_name: str
+    ):
         super().__init__()
         self.host = host
         self.port = port
         self.user = user
         self.password = password
+        self.schema = schema
         self.table_name = table_name
         self.conn = None
         self.postgresql_running: bool = False
         self.postgresql_version: str = ""
         self.number_of_dataproducts: int = 0
+
         self._connect()
         if self.postgresql_running:
             self.postgresql_version = self._get_postgresql_version()
             self.create_metadata_table()
             self.count_jsonb_objects()
+
+    def _check_if_schema_exists(self, schema_name: str) -> bool:
+        """
+        Checks if the specified schema exists in the database.
+
+        Args:
+            schema_name: The name of the schema to check.
+
+        Returns:
+            True if the schema exists, False otherwise.
+        """
+        query = """
+            SELECT count(*)
+            FROM information_schema.schemata
+            WHERE schema_name = %s
+        """
+
+        with self.conn.cursor() as cursor:
+            cursor.execute(query, (schema_name,))
+            return cursor.fetchone()[0] > 0
 
     def status(self) -> dict:
         """
@@ -58,6 +82,7 @@ class PostgresConnector(MetadataStore):
             "port": self.port,
             "user": self.user,
             "running": self.postgresql_running,
+            "schema": self.schema,
             "table_name": self.table_name,
             "number_of_dataproducts": self.number_of_dataproducts,
             "postgresql_version": self.postgresql_version,
@@ -74,8 +99,15 @@ class PostgresConnector(MetadataStore):
             self.conn = psycopg.connect(
                 host=self.host, port=self.port, user=self.user, password=self.password
             )
-            self.postgresql_running = True
-            logger.info("Connected to PostgreSQL successfully")
+            if self._check_if_schema_exists(self.schema):
+                self.postgresql_running = True
+                logger.info("Connected to PostgreSQL successfully")
+            else:
+                self.postgresql_running = False
+                logger.error(
+                    "PostgreSQL schema does not exist, please create schema %s", self.schema
+                )
+
         except OperationalError as error:
             self.postgresql_running = False
             logger.error(
@@ -96,20 +128,31 @@ class PostgresConnector(MetadataStore):
         """
 
         try:
-            logger.info("Creating PostgreSQL metadata table: %s", self.table_name)
+            logger.info(
+                "Creating PostgreSQL metadata table: %s, in schema: %s",
+                self.table_name,
+                self.schema,
+            )
 
-            create_table_query = f"""
-                CREATE TABLE IF NOT EXISTS {self.table_name} (
-                    id SERIAL PRIMARY KEY,
-                    data JSONB NOT NULL,
-                    execution_block VARCHAR(255) DEFAULT NULL UNIQUE,
-                    json_hash CHAR(64) UNIQUE
-                );
-            """
+            # Use parameterization to avoid SQL injection
+            create_table_query = """
+            CREATE TABLE IF NOT EXISTS %(schema_name)s.%(table_name)s (
+                id SERIAL PRIMARY KEY,
+                data JSONB NOT NULL,
+                execution_block VARCHAR(255) DEFAULT NULL UNIQUE,
+                json_hash CHAR(64) UNIQUE
+            );
+            """ % {
+                "schema_name": self.schema,
+                "table_name": self.table_name,
+            }
             cursor = self.conn.cursor()
             cursor.execute(create_table_query)
+            self.conn.commit()
             cursor.close()
-            logger.info("PostgreSQL metadata table %s created.", self.table_name)
+            logger.info(
+                "PostgreSQL metadata table %s created in schema: %s.", self.table_name, self.schema
+            )
 
         except psycopg.Error as error:
             logger.error("Error creating metadata table: %s", error)
@@ -198,7 +241,9 @@ class PostgresConnector(MetadataStore):
     def check_metadata_exists_by_hash(self, json_hash: str) -> bool:
         """Checks if metadata exists based on the given hash."""
         cursor = self.conn.cursor()
-        check_query = f"SELECT EXISTS(SELECT 1 FROM {self.table_name} WHERE json_hash = %s)"
+        check_query = (
+            f"SELECT EXISTS(SELECT 1 FROM {self.schema}.{self.table_name} WHERE json_hash = %s)"
+        )
         cursor.execute(check_query, (json_hash,))
         exists = cursor.fetchone()[0]
         cursor.close()
@@ -207,7 +252,7 @@ class PostgresConnector(MetadataStore):
     def check_metadata_exists_by_execution_block(self, execution_block: str) -> int:
         """Checks if metadata exists based on the given execution block."""
         cursor = self.conn.cursor()
-        check_query = f"SELECT id FROM {self.table_name} WHERE execution_block = %s"
+        check_query = f"SELECT id FROM {self.schema}.{self.table_name} WHERE execution_block = %s"
         cursor.execute(check_query, (execution_block,))
         result = cursor.fetchone()
         cursor.close()
@@ -217,7 +262,9 @@ class PostgresConnector(MetadataStore):
         """Updates existing metadata with the given data and hash."""
         json_hash = self.calculate_metadata_hash(metadata_file_json)
         cursor = self.conn.cursor()
-        update_query = f"UPDATE {self.table_name} SET data = %s, json_hash = %s WHERE id = %s"
+        update_query = (
+            f"UPDATE {self.schema}.{self.table_name} SET data = %s, json_hash = %s WHERE id = %s"
+        )
         cursor.execute(update_query, (metadata_file_json, json_hash, id_field))
         self.conn.commit()
         cursor.close()
@@ -226,8 +273,10 @@ class PostgresConnector(MetadataStore):
         """Inserts new metadata into the database."""
         json_hash = self.calculate_metadata_hash(metadata_file_json)
         cursor = self.conn.cursor()
-        insert_query = f"INSERT INTO {self.table_name} (data, json_hash, execution_block) \
-VALUES (%s, %s, %s)"
+        table: str = self.schema + "." + self.table_name
+        insert_query = (
+            f"INSERT INTO {table} (data, json_hash, execution_block) VALUES (%s, %s, %s)"
+        )
         cursor.execute(insert_query, (metadata_file_json, json_hash, execution_block))
         self.conn.commit()
         cursor.close()
@@ -271,7 +320,7 @@ VALUES (%s, %s, %s)"
         """
 
         cursor = self.conn.cursor()
-        cursor.execute(f"SELECT COUNT(*) FROM {self.table_name}")
+        cursor.execute(f"SELECT COUNT(*) FROM {self.schema}.{self.table_name}")
         result = cursor.fetchone()[0]
         cursor.close()
         self.number_of_dataproducts = int(result)
@@ -290,7 +339,7 @@ VALUES (%s, %s, %s)"
         try:
             logger.info("PostgreSQL deleting database table %s", self.table_name)
             cursor = self.conn.cursor()
-            cursor.execute(f"DROP TABLE IF EXISTS {self.table_name}")
+            cursor.execute(f"DROP TABLE IF EXISTS {self.schema}.{self.table_name}")
             self.conn.commit()
             cursor.close()
             logger.info("PostgreSQL database table %s deleted.", self.table_name)
@@ -315,7 +364,7 @@ VALUES (%s, %s, %s)"
             list[dict]: list of JSON objects.
         """
         cursor = self.conn.cursor()
-        cursor.execute(f"SELECT id, data FROM {table_name}")
+        cursor.execute(f"SELECT id, data FROM {self.schema}.{table_name}")
         data = cursor.fetchall()
         cursor.close()
         return [{"id": row[0], "data": row[1]} for row in data]
@@ -338,12 +387,16 @@ VALUES (%s, %s, %s)"
             A dictionary containing the metadata for the execution block, or None if not found.
         """
         try:
-            return self.get_data_by_execution_block(execution_block)
+            data_product_metadata = self.get_data_by_execution_block(execution_block)
+            if data_product_metadata:
+                return data_product_metadata
+            return {}
+
         except KeyError:
             logger.warning("Metadata not found for execution block: %s", execution_block)
             return {}
 
-    def get_data_by_execution_block(self, execution_block: str) -> dict[str, Any]:
+    def get_data_by_execution_block(self, execution_block: str) -> dict[str, Any] | None:
         """Retrieves data from the PostgreSQL table based on the execution_block.
 
         Args:
@@ -352,15 +405,23 @@ VALUES (%s, %s, %s)"
         Returns:
             The data (JSONB) associated with the execution block, or None if not found.
         """
-        cursor = self.conn.cursor()
-        check_query = f"SELECT data FROM {self.table_name} WHERE execution_block = %s"
-        cursor.execute(check_query, (execution_block,))
-        result = cursor.fetchone()
-        cursor.close()
-        metadata_dict = result[0]
-        if metadata_dict:
-            return metadata_dict
-        return {}
+        try:
+            cursor = self.conn.cursor()
+            check_query = (
+                f"SELECT data FROM {self.schema}.{self.table_name} WHERE execution_block = %s"
+            )
+            cursor.execute(check_query, (execution_block,))
+            result = cursor.fetchone()
+            cursor.close()
+
+            if result:
+                return result[0]
+            return None
+        except Exception as exception:
+            logger.exception(
+                "Error fetching data for execution_block %s: %s", execution_block, exception
+            )
+            raise exception
 
     def get_data_product_file_path(self, execution_block: str) -> pathlib.Path:
         """Retrieves the file path to the data product for the given execution block.
@@ -374,7 +435,9 @@ VALUES (%s, %s, %s)"
 
         try:
             data_product_metadata = self.get_data_by_execution_block(execution_block)
-            return pathlib.Path(data_product_metadata["dataproduct_file"])
+            if data_product_metadata:
+                return pathlib.Path(data_product_metadata["dataproduct_file"])
+            return {}
         except KeyError:
             logger.warning("File path not found for execution block: %s", execution_block)
             return {}